added a few more drawImage function overloads
* The existing drawImage function determined the image type by looking
at the file ending. However, that only works in case the image that's going
to be displayed has a file ending. This changeset adds the possibility
to explicitly specify the format (BMP, JPG, PNG).

* Added the possibility to specify the position (Center, TopLeft,
BottomLeft, TopRight, BottomRight) of the image on the screen. This
function overload is purely for convenience and makes it easier to
position the image on the screen without needing to specify the x and y
pos. Currently, only jpg images fetched from web can be positioned that
way; for every other image format and source (e.g SD card),
the appropriate functionality still needs to be implemented.