Added support for Inkplates that are missing second MCP
New Inkplates are shipped without second MCP I/O expander because of chip shortage. If Inkplate without second MCP I/O expander is programmed using this library, it will hang out on function that is setting it's pins to LOW state, so in this commit is added part of code which stops it.